CDA-98 - Updates packaging of forecast versions. Refactors route configuration into utility class.
