+++ cwms-data-api/src/main/java/cwms/cda/data/dao/forecast/ForecastInstanceDao.java
@@ -1,9 +1,12 @@
-package cwms.cda.data.dao;
+package cwms.cda.data.dao.forecast;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import cwms.cda.api.Controllers;
 import cwms.cda.api.errors.NotFoundException;
+import cwms.cda.data.dao.BlobDao;
+import cwms.cda.data.dao.JooqDao;
+import cwms.cda.data.dao.StreamConsumer;
 import cwms.cda.data.dto.forecast.ForecastInstance;
 import cwms.cda.data.dto.forecast.ForecastSpec;
 import cwms.cda.formatters.json.JsonV2;
