+++ cwms-data-api/src/test/java/cwms/cda/api/ForecastInstanceControllerTestIT.java
@@ -14,7 +14,6 @@ import org.apache.http.NameValuePair;
 import org.apache.http.client.utils.URIBuilder;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
@@ -31,7 +30,7 @@ import java.util.Random;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static cwms.cda.api.ForecastSpecControllerTestIT.*;
+import static cwms.cda.api.ForecastSpecControllerV1TestIT.*;
 import static cwms.cda.security.ApiKeyIdentityProvider.AUTH_HEADER;
 import static io.restassured.RestAssured.given;
 import static java.util.stream.Collectors.toMap;
