+++ cwms-data-api/src/test/resources/cwms/cda/data/dto/forecast/forecast_spec_v2_test.json
